refactor: rename fetcher package to component
We use the (single) word "component" to refer to bosh releases
in a tile. This name matches the team language better.

"*er" named packages are better to avoid because that
naming convention is better to use for interfaces

Authored-by: Christopher Hunter <hunterch@vmware.com>
