+++ internal/commands/update_stemcell_test.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/pivotal-cf/kiln/internal/commands"
 	"github.com/pivotal-cf/kiln/internal/commands/fakes"
-	fetcherFakes "github.com/pivotal-cf/kiln/internal/fetcher/fakes"
+	fetcherFakes "github.com/pivotal-cf/kiln/internal/component/fakes"
 	"github.com/pivotal-cf/kiln/pkg/cargo"
 	"github.com/pivotal-cf/kiln/pkg/release"
 )
