+++ internal/component/bosh_io_release_source.go
@@ -1,4 +1,4 @@
-package fetcher
+package component
 
 import (
 	"crypto/sha1"
