+++ internal/component/bosh_io_release_source_test.go
@@ -1,4 +1,4 @@
-package fetcher_test
+package component_test
 
 import (
 	"crypto/sha1"
@@ -20,19 +20,19 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
-	"github.com/pivotal-cf/kiln/internal/fetcher"
+	"github.com/pivotal-cf/kiln/internal/component"
 	"github.com/pivotal-cf/kiln/pkg/release"
 )
 
 var _ = Describe("BOSHIOReleaseSource", func() {
 	const (
-		ID = fetcher.ReleaseSourceTypeBOSHIO
+		ID = component.ReleaseSourceTypeBOSHIO
 	)
 
 	Describe("GetMatchedReleases from bosh.io", func() {
 		Context("happy path", func() {
 			var (
-				releaseSource *fetcher.BOSHIOReleaseSource
+				releaseSource *component.BOSHIOReleaseSource
 				testServer    *ghttp.Server
 			)
 
@@ -52,7 +52,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				path, _ = regexp.Compile(`/api/v1/releases/github.com/\S+/metrics.*`)
 				testServer.RouteToHandler("GET", path, ghttp.RespondWith(http.StatusOK, `[{"version": "2.3.0"}]`))
 
-				releaseSource = fetcher.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
+				releaseSource = component.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
 			})
 
 			AfterEach(func() {
@@ -69,13 +69,13 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				Expect(err).NotTo(HaveOccurred())
 				Expect(found).To(BeTrue())
 				uaaURL := fmt.Sprintf("%s/d/github.com/cloudfoundry/uaa-release?v=73.3.0", testServer.URL())
-				Expect(foundRelease).To(Equal(release.Remote{ID: release.ID{Name: "uaa", Version: "73.3.0"}, RemotePath: uaaURL, SourceID: fetcher.ReleaseSourceTypeBOSHIO}))
+				Expect(foundRelease).To(Equal(release.Remote{ID: release.ID{Name: "uaa", Version: "73.3.0"}, RemotePath: uaaURL, SourceID: component.ReleaseSourceTypeBOSHIO}))
 
 				foundRelease, found, err = releaseSource.GetMatchedRelease(rabbitmqRequirement)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(found).To(BeTrue())
 				cfRabbitURL := fmt.Sprintf("%s/d/github.com/pivotal-cf/cf-rabbitmq-release?v=268.0.0", testServer.URL())
-				Expect(foundRelease).To(Equal(release.Remote{ID: release.ID{Name: "cf-rabbitmq", Version: "268.0.0"}, RemotePath: cfRabbitURL, SourceID: fetcher.ReleaseSourceTypeBOSHIO}))
+				Expect(foundRelease).To(Equal(release.Remote{ID: release.ID{Name: "cf-rabbitmq", Version: "268.0.0"}, RemotePath: cfRabbitURL, SourceID: component.ReleaseSourceTypeBOSHIO}))
 
 			})
 
@@ -84,7 +84,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 		When("a bosh release doesn't exist on bosh.io in any version", func() {
 			var (
 				testServer    *ghttp.Server
-				releaseSource *fetcher.BOSHIOReleaseSource
+				releaseSource *component.BOSHIOReleaseSource
 			)
 
 			BeforeEach(func() {
@@ -94,7 +94,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				path, _ := regexp.Compile(`/api/v1/releases/github.com/\S+/zzz.*`)
 				testServer.RouteToHandler("GET", path, ghttp.RespondWith(http.StatusOK, `null`))
 
-				releaseSource = fetcher.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
+				releaseSource = component.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
 			})
 
 			AfterEach(func() {
@@ -114,7 +114,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				testServer     *ghttp.Server
 				releaseName    = "my-release"
 				releaseVersion = "1.2.3"
-				releaseSource  *fetcher.BOSHIOReleaseSource
+				releaseSource  *component.BOSHIOReleaseSource
 			)
 
 			BeforeEach(func() {
@@ -123,7 +123,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				pathRegex, _ := regexp.Compile(`/api/v1/releases/github.com/\S+/.*`)
 				testServer.RouteToHandler("GET", pathRegex, ghttp.RespondWith(http.StatusOK, `[{"version": "4.0.4"}]`))
 
-				releaseSource = fetcher.NewBOSHIOReleaseSource(ID, false, testServer.URL(), log.New(GinkgoWriter, "", 0))
+				releaseSource = component.NewBOSHIOReleaseSource(ID, false, testServer.URL(), log.New(GinkgoWriter, "", 0))
 
 			})
 
@@ -149,13 +149,13 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				testServer     *ghttp.Server
 				releaseName    = "my-release"
 				releaseVersion = "1.2.3"
-				releaseSource  *fetcher.BOSHIOReleaseSource
+				releaseSource  *component.BOSHIOReleaseSource
 			)
 
 			BeforeEach(func() {
 				testServer = ghttp.NewServer()
 
-				releaseSource = fetcher.NewBOSHIOReleaseSource(ID, false, testServer.URL(), log.New(GinkgoWriter, "", 0))
+				releaseSource = component.NewBOSHIOReleaseSource(ID, false, testServer.URL(), log.New(GinkgoWriter, "", 0))
 			})
 
 			AfterEach(func() {
@@ -191,7 +191,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 						releaseVersion,
 					)
 
-					Expect(foundRelease).To(Equal(release.Remote{ID: releaseID, RemotePath: expectedPath, SourceID: fetcher.ReleaseSourceTypeBOSHIO}))
+					Expect(foundRelease).To(Equal(release.Remote{ID: releaseID, RemotePath: expectedPath, SourceID: component.ReleaseSourceTypeBOSHIO}))
 				},
 
 				Entry("cloudfoundry org, no suffix", "cloudfoundry", ""),
@@ -218,7 +218,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 		)
 		var (
 			releaseDir    string
-			releaseSource *fetcher.BOSHIOReleaseSource
+			releaseSource *component.BOSHIOReleaseSource
 			testServer    *ghttp.Server
 
 			release1ID release.ID
@@ -234,10 +234,10 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 
 			testServer = ghttp.NewServer()
 
-			releaseSource = fetcher.NewBOSHIOReleaseSource(ID, false, testServer.URL(), log.New(GinkgoWriter, "", 0))
+			releaseSource = component.NewBOSHIOReleaseSource(ID, false, testServer.URL(), log.New(GinkgoWriter, "", 0))
 
 			release1ID = release.ID{Name: "some", Version: "1.2.3"}
-			release1 = release.Remote{ID: release1ID, RemotePath: testServer.URL() + release1ServerPath, SourceID: fetcher.ReleaseSourceTypeBOSHIO}
+			release1 = release.Remote{ID: release1ID, RemotePath: testServer.URL() + release1ServerPath, SourceID: component.ReleaseSourceTypeBOSHIO}
 
 			hash := sha1.New()
 			_, err = io.Copy(hash, strings.NewReader(release1ServerFileContents))
@@ -275,7 +275,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 
 	Describe("FindReleaseVersion from bosh.io", func() {
 		var (
-			releaseSource *fetcher.BOSHIOReleaseSource
+			releaseSource *component.BOSHIOReleaseSource
 			testServer    *ghttp.Server
 		)
 		When("a bosh release exist on bosh.io", func() {
@@ -286,7 +286,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				path, _ := regexp.Compile(`/api/v1/releases/github.com/\S+/cf-rabbitmq.*`)
 				testServer.RouteToHandler("GET", path, ghttp.RespondWith(http.StatusOK, `[{"name":"github.com/cloudfoundry/cf-rabbitmq-release","version":"309.0.5","url":"https://bosh.io/d/github.com/cloudfoundry/cf-rabbitmq-release?v=309.0.0","sha1":"5df538657c2cc830bda679420a9b162682018ded"},{"name":"github.com/cloudfoundry/cf-rabbitmq-release","version":"308.0.0","url":"https://bosh.io/d/github.com/cloudfoundry/cf-rabbitmq-release?v=308.0.0","sha1":"56202c9a466a8394683ae432ee2dea21ef6ef865"}]`))
 
-				releaseSource = fetcher.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
+				releaseSource = component.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
 			})
 
 			AfterEach(func() {
@@ -300,7 +300,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 					Expect(err).NotTo(HaveOccurred())
 					Expect(found).To(BeTrue())
 					cfRabbitURL := fmt.Sprintf("%s/d/github.com/cloudfoundry/cf-rabbitmq-release?v=309.0.5", testServer.URL())
-					Expect(foundRelease).To(Equal(release.Remote{ID: release.ID{Name: "cf-rabbitmq", Version: "309.0.5"}, SHA: "5df538657c2cc830bda679420a9b162682018ded", RemotePath: cfRabbitURL, SourceID: fetcher.ReleaseSourceTypeBOSHIO}))
+					Expect(foundRelease).To(Equal(release.Remote{ID: release.ID{Name: "cf-rabbitmq", Version: "309.0.5"}, SHA: "5df538657c2cc830bda679420a9b162682018ded", RemotePath: cfRabbitURL, SourceID: component.ReleaseSourceTypeBOSHIO}))
 
 				})
 			})
@@ -315,7 +315,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 					Expect(foundRelease).To(Equal(release.Remote{ID: release.ID{Name: "cf-rabbitmq", Version: "309.0.5"},
 						SHA:        "5df538657c2cc830bda679420a9b162682018ded",
 						RemotePath: cfRabbitURL,
-						SourceID:   fetcher.ReleaseSourceTypeBOSHIO}))
+						SourceID:   component.ReleaseSourceTypeBOSHIO}))
 				})
 			})
 		})
@@ -327,7 +327,7 @@ var _ = Describe("BOSHIOReleaseSource", func() {
 				path, _ := regexp.Compile(`/api/v1/releases/github.com/\S+/cf-rabbitmq.*`)
 				testServer.RouteToHandler("GET", path, ghttp.RespondWith(http.StatusOK, `null`))
 
-				releaseSource = fetcher.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
+				releaseSource = component.NewBOSHIOReleaseSource(ID, false, testServer.URL(), logger)
 			})
 
 			AfterEach(func() {
