+++ internal/component/errors.go
@@ -1,4 +1,4 @@
-package fetcher
+package component
 
 type stringError string
