+++ internal/component/fakes/read_closer.go
