+++ internal/component/fakes/remote_pather.go
@@ -4,7 +4,7 @@ package fakes
 import (
 	"sync"
 
-	"github.com/pivotal-cf/kiln/internal/fetcher"
+	"github.com/pivotal-cf/kiln/internal/component"
 	"github.com/pivotal-cf/kiln/pkg/release"
 )
 
@@ -114,4 +114,4 @@ func (fake *RemotePather) recordInvocation(key string, args []interface{}) {
 	fake.invocations[key] = append(fake.invocations[key], args)
 }
 
-var _ fetcher.RemotePather = new(RemotePather)
+var _ component.RemotePather = new(RemotePather)
