+++ internal/component/fakes/s3_client.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/aws/aws-sdk-go/service/s3"
-	"github.com/pivotal-cf/kiln/internal/fetcher"
+	"github.com/pivotal-cf/kiln/internal/component"
 )
 
 type S3Client struct {
@@ -193,4 +193,4 @@ func (fake *S3Client) recordInvocation(key string, args []interface{}) {
 	fake.invocations[key] = append(fake.invocations[key], args)
 }
 
-var _ fetcher.S3Client = new(S3Client)
+var _ component.S3Client = new(S3Client)
