+++ internal/component/fakes/s3_request_failure.go
