+++ internal/component/fakes/s3_uploader.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
-	"github.com/pivotal-cf/kiln/internal/fetcher"
+	"github.com/pivotal-cf/kiln/internal/component"
 )
 
 type S3Uploader struct {
@@ -116,4 +116,4 @@ func (fake *S3Uploader) recordInvocation(key string, args []interface{}) {
 	fake.invocations[key] = append(fake.invocations[key], args)
 }
 
-var _ fetcher.S3Uploader = new(S3Uploader)
+var _ component.S3Uploader = new(S3Uploader)
