+++ internal/component/fetcher_suite_test.go
@@ -1,4 +1,4 @@
-package fetcher_test
+package component_test
 
 import (
 	"testing"
