+++ internal/component/local_release_directory.go
@@ -1,4 +1,4 @@
-package fetcher
+package component
 
 import (
 	"crypto/sha1"
